+++ sdk/java/src/main/java/com/pulumi/eks/NodeGroupArgs.java
@@ -6,9 +6,11 @@ package com.pulumi.eks;
 import com.pulumi.aws.ec2.SecurityGroup;
 import com.pulumi.aws.ec2.SecurityGroupRule;
 import com.pulumi.aws.iam.InstanceProfile;
+import com.pulumi.core.Either;
 import com.pulumi.core.Output;
 import com.pulumi.core.annotations.Import;
 import com.pulumi.eks.Cluster;
+import com.pulumi.eks.inputs.CoreDataArgs;
 import com.pulumi.eks.inputs.TaintArgs;
 import java.lang.Boolean;
 import java.lang.Integer;
@@ -138,13 +140,13 @@ public final class NodeGroupArgs extends com.pulumi.resources.ResourceArgs {
      * 
      */
     @Import(name="cluster", required=true)
-    private Cluster cluster;
+    private Output<Either<Cluster,CoreDataArgs>> cluster;
 
     /**
      * @return The target EKS cluster.
      * 
      */
-    public Cluster cluster() {
+    public Output<Either<Cluster,CoreDataArgs>> cluster() {
         return this.cluster;
     }
 
@@ -720,9 +722,39 @@ public final class NodeGroupArgs extends com.pulumi.resources.ResourceArgs {
          * @return builder
          * 
          */
-        public Builder cluster(Cluster cluster) {
+        public Builder cluster(Output<Either<Cluster,CoreDataArgs>> cluster) {
             $.cluster = cluster;
             return this;
+        }
+
+        /**
+         * @param cluster The target EKS cluster.
+         * 
+         * @return builder
+         * 
+         */
+        public Builder cluster(Either<Cluster,CoreDataArgs> cluster) {
+            return cluster(Output.of(cluster));
+        }
+
+        /**
+         * @param cluster The target EKS cluster.
+         * 
+         * @return builder
+         * 
+         */
+        public Builder cluster(Cluster cluster) {
+            return cluster(Either.ofLeft(cluster));
+        }
+
+        /**
+         * @param cluster The target EKS cluster.
+         * 
+         * @return builder
+         * 
+         */
+        public Builder cluster(CoreDataArgs cluster) {
+            return cluster(Either.ofRight(cluster));
         }
 
         /**